refactor(configuration): Move Socket into the Directive directory, Configuration is now a struct
