+++ src/Configuration.cpp
@@ -1,0 +1,7 @@
+#include "Configuration.hpp"
+
+#include <cassert>
+
+#include <vector>
+
+Configuration ws_configuration;
